added unfocus support
